+++ tests/integration/main.rs
@@ -49,7 +49,8 @@ async fn test_added_notes() {
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     println!("Running Mint tx...");
     execute_tx_and_sync(&mut client, faucet_account_header.id(), tx_request).await;
 
@@ -85,7 +86,8 @@ async fn test_multiple_tx_on_same_block() {
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     let tx_request_2 = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         None,
@@ -93,7 +95,8 @@ async fn test_multiple_tx_on_same_block() {
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
 
     println!("Running P2ID tx...");
 
@@ -177,7 +180,8 @@ async fn test_p2id_transfer() {
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
 
     let note = tx_request.expected_output_notes().next().unwrap().clone();
     let transaction_id = execute_tx(&mut client, from_account_id, tx_request).await;
@@ -207,7 +211,7 @@ async fn test_p2id_transfer() {
 
     // Consume P2ID note
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build();
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build().unwrap();
     execute_tx_and_sync(&mut client, to_account_id, tx_request).await;
 
     // Ensure we have nothing else to consume
@@ -270,7 +274,8 @@ async fn test_p2id_transfer_failing_not_enough_balance() {
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     execute_failing_tx(
         &mut client,
         from_account_id,
@@ -352,7 +357,8 @@ async fn test_p2idr_transfer_consumed_by_target() {
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     execute_tx_and_sync(&mut client, from_account_id, tx_request.clone()).await;
 
     // Check that note is committed for the second account to consume
@@ -363,7 +369,7 @@ async fn test_p2idr_transfer_consumed_by_target() {
     // Make the `to_account_id` consume P2IDR note
     let note_id = tx_request.expected_output_notes().next().unwrap().id();
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_id]).build();
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_id]).build().unwrap();
     execute_tx_and_sync(&mut client, to_account_id, tx_request).await;
     let regular_account = client.get_account(from_account_id).await.unwrap().unwrap();
 
@@ -430,7 +436,8 @@ async fn test_p2idr_transfer_consumed_by_sender() {
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
 
     // Check that note is committed
@@ -440,7 +447,7 @@ async fn test_p2idr_transfer_consumed_by_sender() {
 
     // Check that it's still too early to consume
     println!("Consuming Note (too early)...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build();
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build().unwrap();
     let transaction_execution_result = client.new_transaction(from_account_id, tx_request).await;
     assert!(transaction_execution_result.is_err_and(|err| {
         matches!(
@@ -461,7 +468,7 @@ async fn test_p2idr_transfer_consumed_by_sender() {
 
     // Consume the note with the sender account
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build();
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build().unwrap();
     execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
 
     let regular_account = client.get_account(from_account_id).await.unwrap().unwrap();
@@ -521,7 +528,8 @@ async fn test_get_consumable_notes() {
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
 
     // Check that note is consumable by both accounts
@@ -589,7 +597,8 @@ async fn test_get_output_notes() {
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
 
     let output_note_id = tx_request.expected_output_notes().next().unwrap().id();
 
@@ -624,7 +633,8 @@ async fn test_import_expected_notes() {
         client_2.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
     client_2.sync_state().await.unwrap();
 
@@ -658,7 +668,8 @@ async fn test_import_expected_notes() {
         client_2.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
 
     // Import an uncommited note without verification
@@ -709,7 +720,8 @@ async fn test_import_expected_note_uncommitted() {
         client_1.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
 
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
     client_2.sync_state().await.unwrap();
@@ -750,7 +762,8 @@ async fn test_import_expected_notes_from_the_past_as_committed() {
         client_1.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
 
     let block_height_before = client_1.get_sync_height().await.unwrap();
@@ -856,7 +869,8 @@ async fn test_sync_detail_values() {
         client1.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     let note_id = tx_request.expected_output_notes().next().unwrap().id();
     execute_tx_and_sync(&mut client1, from_account_id, tx_request).await;
 
@@ -868,7 +882,7 @@ async fn test_sync_detail_values() {
     assert_eq!(new_details.updated_accounts.len(), 0);
 
     // Consume the note with the second account
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_id]).build();
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_id]).build().unwrap();
     execute_tx_and_sync(&mut client2, to_account_id, tx_request).await;
 
     // First client sync should have a new nullifier as the note was consumed
@@ -903,7 +917,8 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
             client.rng(),
         )
         .unwrap()
-        .build();
+        .build()
+        .unwrap();
 
         println!("Executing transaction...");
         let transaction_execution_result =
@@ -930,7 +945,8 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
             client.rng(),
         )
         .unwrap()
-        .build();
+        .build()
+        .unwrap();
 
         println!("Executing transaction...");
         let transaction_execution_result =
@@ -967,7 +983,8 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
             client.rng(),
         )
         .unwrap()
-        .build();
+        .build()
+        .unwrap();
 
         println!("Executing transaction...");
         let transaction_execution_result =
@@ -1076,13 +1093,15 @@ async fn test_consume_multiple_expected_notes() {
         client_owned_notes.iter().map(|note| note.id()).collect(),
     )
     .with_authenticated_input_notes(client_owned_notes.iter().map(|note| (note.id(), None)))
-    .build();
+    .build()
+    .unwrap();
 
     let tx_request_2 = TransactionRequestBuilder::consume_notes(
         unauth_owned_notes.iter().map(|note| note.id()).collect(),
     )
     .with_unauthenticated_input_notes(unauth_owned_notes.iter().map(|note| ((*note).clone(), None)))
-    .build();
+    .build()
+    .unwrap();
 
     let tx_id_1 = execute_tx(&mut client, to_account_ids[0], tx_request_1).await;
     let tx_id_2 = execute_tx(&mut unauth_client, to_account_ids[1], tx_request_2).await;
@@ -1155,7 +1174,8 @@ async fn test_import_consumed_note_with_proof() {
         client_1.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     let note = client_1
         .get_input_notes(NoteFilter::Committed)
@@ -1168,7 +1188,7 @@ async fn test_import_consumed_note_with_proof() {
     // Consume the note with the sender account
 
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build();
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build().unwrap();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
 
     // Import the consumed note
@@ -1214,7 +1234,8 @@ async fn test_import_consumed_note_with_id() {
         client_1.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     let note = client_1
         .get_input_notes(NoteFilter::Committed)
@@ -1227,7 +1248,7 @@ async fn test_import_consumed_note_with_id() {
     // Consume the note with the sender account
 
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build();
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build().unwrap();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     client_2.sync_state().await.unwrap();
 
@@ -1272,7 +1293,8 @@ async fn test_discarded_transaction() {
         client_1.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
 
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     client_2.sync_state().await.unwrap();
@@ -1285,7 +1307,7 @@ async fn test_discarded_transaction() {
         .clone();
 
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build();
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build().unwrap();
 
     // Consume the note in client 1 but dont submit it to the node
     let tx_result = client_1.new_transaction(from_account_id, tx_request.clone()).await.unwrap();
@@ -1353,7 +1375,8 @@ async fn test_custom_transaction_prover() {
         client.rng(),
     )
     .unwrap()
-    .build();
+    .build()
+    .unwrap();
 
     let transaction_execution_result =
         client.new_transaction(faucet_account_id, tx_request.clone()).await.unwrap();
@@ -1464,8 +1487,8 @@ async fn test_expired_transaction_fails() {
     )
     .unwrap()
     .with_expiration_delta(expiration_delta)
-    .unwrap()
-    .build();
+    .build()
+    .unwrap();
 
     println!("Executing transaction...");
     let transaction_execution_result =